backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
numpy==1.26.4
h5py==3.10.0
